requirements.txt
asgiref==3.9.1
astroid==3.3.11
certifi==2025.7.14
charset-normalizer==3.4.2
dill==0.4.0
Django==5.2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
exceptiongroup==1.3.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
isort==6.0.1
mccabe==0.7.0
packaging==25.0
platformdirs==4.3.8
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.10.1
pylint==3.3.7
pytest==8.4.1
pytest-django==4.11.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
sqlparse==0.5.3
tomli==2.2.1
tomlkit==0.13.3
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
